// src/centers/enum/center-type.enum.ts
export enum CenterType {
  HOSPITAL = 'hospital',
  PHARMACY = 'pharmacy',
  CLINIC = 'clinic',
  LABORATORY = 'laboratory',
  RADIOLOGY = 'radiology',
  DENTAL = 'dental',
  EYE = 'eye',
  MATERNITY = 'maternity',
  AMBULANCE = 'ambulance',
  VIROLOGY = 'virology',
  PSYCHIATRIC = 'psychiatric',
  CARE_HOME = 'care-home',
  HOSPICE = 'hospice',
  FUNERAL = 'funeral'
} 